refactor: extract shared MockEventSource test utility
Move duplicated MockEventSource class from useEventSource.test.ts and
useTimeSeriesStream.test.ts into a shared test-utils module. The
useTimeSeriesStream copy was missing url and closed fields which could
mask bugs in future tests.
