+++ src/hooks/__tests__/useEventSource.test.ts
@@ -1,51 +1,7 @@
 import { describe, it, expect, beforeEach, afterEach, mock } from 'bun:test';
 import { renderHook, act } from '@testing-library/react';
 import { useEventSource } from '../useEventSource';
-
-class MockEventSource {
-  static instances: MockEventSource[] = [];
-  url: string;
-  onopen: (() => void) | null = null;
-  onmessage: ((event: { data: string }) => void) | null = null;
-  onerror: (() => void) | null = null;
-  private listeners = new Map<string, ((event: unknown) => void)[]>();
-  readyState = 0;
-  closed = false;
-
-  static readonly CONNECTING = 0;
-  static readonly OPEN = 1;
-  static readonly CLOSED = 2;
-
-  constructor(url: string) {
-    this.url = url;
-    MockEventSource.instances.push(this);
-  }
-
-  addEventListener(type: string, handler: (event: unknown) => void) {
-    const handlers = this.listeners.get(type) ?? [];
-    handlers.push(handler);
-    this.listeners.set(type, handlers);
-  }
-
-  removeEventListener(type: string, handler: (event: unknown) => void) {
-    const handlers = this.listeners.get(type) ?? [];
-    this.listeners.set(type, handlers.filter(h => h !== handler));
-  }
-
-  close() {
-    this.closed = true;
-    this.readyState = 2;
-  }
-
-  fireEvent(type: string) {
-    const handlers = this.listeners.get(type) ?? [];
-    for (const handler of handlers) handler({});
-  }
-
-  static reset() {
-    MockEventSource.instances = [];
-  }
-}
+import { MockEventSource } from './test-utils/mock-event-source';
 
 const originalEventSource = globalThis.EventSource;
