+++ src/hooks/__tests__/useTimeSeriesStream.test.ts
@@ -1,41 +1,7 @@
 import { describe, it, expect, beforeEach, afterEach, mock } from 'bun:test';
 import { renderHook, act } from '@testing-library/react';
 import { useTimeSeriesStream, VISIBILITY_REFRESH_COOLDOWN_MS } from '../useTimeSeriesStream';
-
-// MockEventSource with listener support for stats_error dispatch
-class MockEventSource {
-  static instances: MockEventSource[] = [];
-  onopen: (() => void) | null = null;
-  onmessage: ((event: { data: string }) => void) | null = null;
-  onerror: (() => void) | null = null;
-  private listeners = new Map<string, ((event: unknown) => void)[]>();
-  readyState = 0;
-
-  static readonly CONNECTING = 0;
-  static readonly OPEN = 1;
-  static readonly CLOSED = 2;
-
-  constructor(_url: string) { MockEventSource.instances.push(this); }
-
-  addEventListener(type: string, handler: (event: unknown) => void) {
-    const handlers = this.listeners.get(type) ?? [];
-    handlers.push(handler);
-    this.listeners.set(type, handlers);
-  }
-
-  removeEventListener(type: string, handler: (event: unknown) => void) {
-    const handlers = this.listeners.get(type) ?? [];
-    this.listeners.set(type, handlers.filter(h => h !== handler));
-  }
-
-  fireEvent(type: string) {
-    const handlers = this.listeners.get(type) ?? [];
-    for (const h of handlers) h({});
-  }
-
-  close() { this.readyState = 2; }
-  static reset() { MockEventSource.instances = []; }
-}
+import { MockEventSource } from './test-utils/mock-event-source';
 
 const originalEventSource = globalThis.EventSource;
